requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
wandb
albumentations
gdown
Pillow